+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+- rename `Decodable.init(from:)` to `Decodable.init(dictionary:)` to avoid clash with R.swift .
+
 ## [1.1.2] - 2020-02-07
 
 ### Fixed